WebSockets/tests/submissions/Opera/cookies/005.html

<!doctype html>
<title>WebSockets: setting HttpOnly cookies in ws response, checking ws request</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.js></script>
<div id=log></div>
<script>

var timeout = setTimeout(function() {
  deleteCookie(false, 'timed out');
}, 15000);
var iframe = document.createElement('iframe');
try {
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/ws/set-cookie?'+encodeURIComponent('ws_test=test; Path=/; HttpOnly'));
  ws.onopen = function(e) {
    
    ws.onclose = debug;
    var ws2 = new WebSocket(SCHEME_DOMAIN_PORT+'/ws/echo-cookie');
    ws2.onmessage = function(e) {
      debug('ws2 message');
      ws2.onclose = function(e) { debug('ws2 close'); }
      ws.close();
      ws2.close();
      deleteCookie(/ws_test=test/.test(e.data), e.data);
    }
    ws2.onopen = function(e) { debug('ws2 open'); }
    ws2.onclose = function(e) { debug('ws2 close'); deleteCookie(false, 'ws2 close'); }
  }
  ws.onclose = t.step_func(function() {assert_unreached()});
} catch(e) {
  assert_unreached(e);
}

function deleteCookie(testPassed, msg) {
  if (arguments.callee.done)
    return;
  arguments.callee.done = true;
  clearTimeout(timeout);
  setTestTimeout(10000);
  iframe.onload = function() {
    assert_equals(testPassed, true, msg);
    end();
  }
  // remove cookie
  iframe.src = 'support/set-cookie.php?'+encodeURIComponent('ws_test=; Path=/; HttpOnly; Expires=Sun, 06 Nov 1994 08:49:37 GMT');
}
document.body.appendChild(iframe);
</script>
